Upgrade of mobility_features packages
