+++ packages/mobility_features/example/pubspec.yaml
@@ -3,15 +3,15 @@ description: An example app for the Mobility Features package.
 publish_to: "none" 
 
 environment:
-  sdk: ">=3.2.0 <4.0.0"
-  flutter: ">=3.16.0"
+  sdk: ">=3.8.0 <4.0.0"
+  flutter: ">=3.19.0"
 
 dependencies:
   flutter:
     sdk: flutter
   permission_handler: ^11.3.1
   carp_background_location: ^4.0.0
-
+  
   mobility_features:
     path: ../
