+++ packages/mobility_features/pubspec.yaml
@@ -1,17 +1,17 @@
 name: mobility_features
 description: Calculation of real-time mobility features like places, stops, and home stay
-version: 6.0.0
+version: 6.1.0
 homepage: https://github.com/cph-cachet/flutter-plugins/
 
 environment:
-  sdk: ">=3.2.0 <4.0.0"
-  flutter: ">=3.16.0"
+  sdk: ">=3.8.0 <4.0.0"
+  flutter: ">=3.19.0"
 
 dependencies:
   flutter:
     sdk: flutter
   simple_cluster: ^0.3.0
-  stats: ^2.0.0
+  stats: ^3.0.0
   path_provider: ^2.0.2
   json_annotation: ^4.8.0
   carp_serializable: ^2.0.0
